Agregue controlador checador y incluid con js en stripe
